Updated Product To Retrieve Right Data Type
getProductQuantity was returning Double and not long

+++ SpringRestCRM/src/main/java/com/lorenzo/models/Product.java
@@ -42,7 +42,7 @@ public class Product {
 	public void setProductPrice(double productPrice) {
 		this.productPrice = productPrice;
 	}
-	public double getProductQuantity() {
+	public long getProductQuantity() {
 		return productQuantity;
 	}
 	public void setProductQuantity(long productQuantity) {
